Move SelectedPortWaitStrategy to trino-testing-containers
