+++ testing/trino-product-tests-launcher/src/main/java/io/trino/tests/product/launcher/docker/ContainerUtil.java
@@ -20,7 +20,7 @@ import com.github.dockerjava.api.exception.ConflictException;
 import com.github.dockerjava.api.exception.NotFoundException;
 import com.github.dockerjava.api.model.Container;
 import com.github.dockerjava.api.model.Network;
-import io.trino.tests.product.launcher.testcontainers.SelectedPortWaitStrategy;
+import io.trino.testing.containers.wait.strategy.SelectedPortWaitStrategy;
 import org.testcontainers.DockerClientFactory;
 import org.testcontainers.containers.wait.strategy.WaitStrategy;
