+++ testing/trino-testing-containers/pom.xml
@@ -39,6 +39,12 @@
         </dependency>
 
         <dependency>
+            <groupId>org.rnorth.duct-tape</groupId>
+            <artifactId>duct-tape</artifactId>
+            <version>${dep.duct-tape.version}</version>
+        </dependency>
+
+        <dependency>
             <groupId>org.testcontainers</groupId>
             <artifactId>testcontainers</artifactId>
         </dependency>
